Moved TypeValidator, TypesValidator to message.py
